Add plugins and tools page.
Add plugins and tools page.  This will require feedback for what we want
to do and how to curate this list.  There are currently freemium plugins
in this list which are on the WP.org repo and there are also plugins.
It would be nice to have a consensus of how we will curate a list of
extensions and things built with wp api moving into the future.