Add callouts to new documentation site on all pages that have been ported
![image](https://cloud.githubusercontent.com/assets/442115/22086554/38292a8e-dda7-11e6-9034-841068a6ef4b.png)

This adds a callout like that, with a link to the relevant page, to each
document within this site that has been moved to the developer.wp.org
REST API handbook. This also makes it easier to see what has yet to be
migrated.

The top nav has been adjusted to point to the now-relevant project-wide
support forums and main WordPress download link, to discourage future
use of the beta plugin variant.